+++ get_next_line_bonus.c
@@ -6,7 +6,7 @@
 /*   By: aokubo <aokubo@student.42tokyo.jp>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/06 18:06:50 by aokubo            #+#    #+#             */
-/*   Updated: 2022/03/16 17:44:34 by aokubo           ###   ########.fr       */
+/*   Updated: 2022/04/24 05:04:31 by aokubo           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -71,7 +71,7 @@ ssize_t	ft_read_file_gnl(t_list_gnl *lst_fd, t_list_gnl **lst)
 		str_read = (char *)malloc(sizeof(char) * ((size_t)BUFFER_SIZE + 1));
 		if (str_read != NULL)
 			str_read[0] = '\0';
-		read_size = read(lst_fd->fd, str_read, BUFFER_SIZE);
+		read_size = read(lst_fd->fd, str_read, (size_t)BUFFER_SIZE);
 		if (read_size != -1)
 			str_read[read_size] = '\0';
 		lst_fd->content = ft_strjoin_gnl(lst_fd->content, str_read, true, true);
